+++ crates/bevy_ui/src/widget/mod.rs
@@ -4,12 +4,12 @@ mod button;
 mod image;
 mod label;
 mod text;
-mod text_editable;
+mod text_input_layout;
 mod viewport;
 
 pub use button::*;
 pub use image::*;
 pub use label::*;
 pub use text::*;
-pub use text_editable::*;
+pub use text_input_layout::*;
 pub use viewport::*;
